refactor: extract updating issues
